Extole: renaming subdir
