+++ lib/extole/index.js
@@ -7,8 +7,8 @@
 var bind = require('bind');
 var each = require('each');
 var integration = require('analytics.js-integration');
-var formatConversionParams = require('./lib/formatConversionParams');
-var createConversionTag = require('./lib/createConversionTag');
+var formatConversionParams = require('./extole/formatConversionParams');
+var createConversionTag = require('./extole/createConversionTag');
 
 /**
 * Expose `Extole` integration.
