+++ lib/extole/test.js
@@ -6,7 +6,7 @@ var integration = require('analytics.js-integration');
 var Extole = require('./');
 var sandbox = require('clear-env');
 var tester = require('analytics.js-integration-tester');
-var formatConversionParams = require('./lib/formatConversionParams');
+var formatConversionParams = require('./extole/formatConversionParams');
 
 describe('Extole', function(){
   var extole;
